Use (u)int64_t in place of long long
$ find -type f -exec sed -i 's/unsigned long long/uint64_t/g' {} +
$ find -type f -exec sed -i 's/long long/int64_t/g' {} +

+++ src/libsodium/include/sodium/crypto_auth.h
@@ -24,11 +24,11 @@ const char *crypto_auth_primitive(void);
 
 SODIUM_EXPORT
 int crypto_auth(unsigned char *out, const unsigned char *in,
-                unsigned long long inlen, const unsigned char *k);
+                uint64_t inlen, const unsigned char *k);
 
 SODIUM_EXPORT
 int crypto_auth_verify(const unsigned char *h, const unsigned char *in,
-                       unsigned long long inlen, const unsigned char *k);
+                       uint64_t inlen, const unsigned char *k);
 #ifdef __cplusplus
 }
 #endif
