+++ src/libsodium/include/sodium/crypto_onetimeauth.h
@@ -24,11 +24,11 @@ const char *crypto_onetimeauth_primitive(void);
 
 SODIUM_EXPORT
 int crypto_onetimeauth(unsigned char *out, const unsigned char *in,
-                       unsigned long long inlen, const unsigned char *k);
+                       uint64_t inlen, const unsigned char *k);
 
 SODIUM_EXPORT
 int crypto_onetimeauth_verify(const unsigned char *h, const unsigned char *in,
-                              unsigned long long inlen, const unsigned char *k);
+                              uint64_t inlen, const unsigned char *k);
 
 #ifdef __cplusplus
 }
