+++ src/libsodium/include/sodium/crypto_secretbox.h
@@ -36,12 +36,12 @@ const char *crypto_secretbox_primitive(void);
 
 SODIUM_EXPORT
 int crypto_secretbox(unsigned char *c, const unsigned char *m,
-                     unsigned long long mlen, const unsigned char *n,
+                     uint64_t mlen, const unsigned char *n,
                      const unsigned char *k);
 
 SODIUM_EXPORT
 int crypto_secretbox_open(unsigned char *m, const unsigned char *c,
-                          unsigned long long clen, const unsigned char *n,
+                          uint64_t clen, const unsigned char *n,
                           const unsigned char *k);
 
 #ifdef __cplusplus
